LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-web-socket.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

TIMEOUT WebSocket inherits storage access Test timed out
NOTRUN WebSocket omits unpartitioned cookies without storage access
